Update LLama3 example to be a generic LLM pipleine example, and added some (temporary) memory profing sanity checks
